Modify: master 修改手机号增加状态管理[https://github.com/AndromedaX7/worldfunclub/issues/27]
